exception

+++ IO/Filewriter.java
@@ -1,5 +1,4 @@
 package IO;
-
 import java.io.BufferedWriter;
 import java.io.FileWriter;
 import java.io.IOException;
@@ -10,8 +9,11 @@ import java.nio.file.Paths;
 public class Filewriter {
   
     public String createProcessFile(String filename, String fileContent,String path) throws IOException {
-      Path p=Paths.get(path);    
-        String newFilename = filename.replaceAll(".{5}$", ".txt");
+     Path p=Paths.get(path);
+        if (!Files.exists(dirPath) || !Files.isDirectory(dirPath)) {
+            throw new IllegalArgumentException("Invalid directory path: " + directoryPath);
+        }
+        //String newFilename = filename.replaceAll(".{5}$", ".txt");
         Path newpath = Paths.get(path);
         BufferedWriter br = new BufferedWriter(new FileWriter(path + "//" + newFilename));
         br.write(fileContent);
@@ -20,3 +22,4 @@ public class Filewriter {
     }
 
 }
+
